refactor(certificate): move from page to component content
